Working MCMC Sampler
